requirements.txt
dj-database-url==0.5.0
Django==2.1
django-bootstrap3==10.0.1
django-environ==0.4.5
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
raven==6.9.0
six==1.11.0
whitenoise==4.0
django-redis==4.9.0
requests==2.19.1
rq==0.12.0
pillow==5.2.0
boto3==1.7.80
botocore==1.10.80
django-storages==1.6.6
django-phonenumber-field
beautifulsoup4==4.6.3